+++ contrib/utils/install-dependencies-bookworm.sh
@@ -50,7 +50,6 @@ PACKAGES=(
   libdb-dev
   libevent-dev
   libjemalloc-dev
-  libminiupnpc-dev
   libprotobuf-dev
   libpcsclite-dev
   libqrencode-dev