+++ src/clientversion.cpp
@@ -114,8 +114,7 @@ std::string LicenseInfo() {
            "\n" + "\n" +
            strprintf(_("This product includes software developed by the "
                        "OpenSSL Project for use in the OpenSSL Toolkit %s and "
-                       "cryptographic software written by Eric Young and UPnP "
-                       "software written by Thomas Bernard.")
+                       "cryptographic software written by Eric Young.")
                          .translated,
                      "<https://www.openssl.org>") +
            "\n";