+++ src/mapport.h
@@ -5,18 +5,16 @@
 #ifndef BITCOIN_MAPPORT_H
 #define BITCOIN_MAPPORT_H
 
-static constexpr bool DEFAULT_UPNP = false;
-
 static constexpr bool DEFAULT_NATPMP = false;
 
 enum MapPortProtoFlag : unsigned int {
     NONE = 0x00,
-    UPNP = 0x01,
+    // 0x01 was for UPnP, for which we dropped support.
     // PCP with NAT-PMP fallback.
     PCP = 0x02,
 };
 
-void StartMapPort(bool use_upnp, bool use_pcp);
+void StartMapPort(bool use_pcp);
 void InterruptMapPort();
 void StopMapPort();