+++ src/node/interfaces.cpp
@@ -168,9 +168,7 @@ namespace {
             });
             gArgs.WriteSettingsFile();
         }
-        void mapPort(bool use_upnp, bool use_pcp) override {
-            StartMapPort(use_upnp, use_pcp);
-        }
+        void mapPort(bool use_pcp) override { StartMapPort(use_pcp); }
         bool getProxy(Network net, Proxy &proxy_info) override {
             return GetProxy(net, proxy_info);
         }