+++ src/qt/optionsmodel.h
@@ -56,7 +56,6 @@ public:
         StartAtStartup,      // bool
         HideTrayIcon,        // bool
         MinimizeToTray,      // bool
-        MapPortUPnP,         // bool
         MapPortNatpmp,       // bool
         MinimizeOnClose,     // bool
         ProxyUse,            // bool