+++ doc/release-notes.md
@@ -4,4 +4,7 @@ Bitcoin ABC version 0.33.1 is now available from:
 
   <https://download.bitcoinabc.org/0.33.1/>
 
-This is a maintenance release with no user-visible change.
+This release includes the following features and fixes:
+- Support for UPnP was dropped. If you want to open a port automatically, consider using
+  the `-natpmp` option instead, which uses PCP or NAT-PMP depending on router support.
+  Setting `-upnp` will now return an error.